BugFixes/DeviceTokenJun9

+++ networklayer/src/main/java/com/emproto/networklayer/response/responsee/GeneralInfoEscalationGraph.kt
@@ -2,7 +2,7 @@ package com.emproto.networklayer.response.responsee
 
 data class GeneralInfoEscalationGraph(
     val dataPoints: DataPoints,
-    val estimatedAppreciation: Int,
+    val estimatedAppreciation: Double,
     val xAxisDisplayName: String,
     val yAxisDisplayName: String
 )
